api/requirements.txt
beautifulsoup4
pandas
openai
python-dotenv
pyarrow
scikit-learn
flask
flask-cors
numpy
Gunicorn
pinecone-client
fake-useragent